Move integration test classes to their own files.
