+++ src/test/java/org/example/CassandraSimpleIntegrationTest.java
@@ -2,22 +2,12 @@ package org.example;
 
 import com.datastax.driver.core.Cluster;
 import com.datastax.driver.core.Session;
-import com.datastax.driver.core.utils.UUIDs;
 import com.datastax.oss.driver.api.core.CqlSession;
-import org.example.dao.Car;
-import org.example.dao.CarRepository;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringBootVersion;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.testcontainers.containers.CassandraContainer;
 import org.testcontainers.junit.jupiter.Testcontainers;
-
-import java.util.List;
-import java.util.UUID;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest(classes = {
         SampleCassandraApplication.class
@@ -84,56 +74,6 @@ abstract class CassandraSimpleIntegrationTest {
                     " WITH replication = \n" +
                     "{'class':'SimpleStrategy','replication_factor':'1'};");
         }
-    }
-
-}
-
-class ApplicationContextLiveTest extends CassandraSimpleIntegrationTest {
-
-    @Test
-    void givenCassandraContainer_whenSpringContextIsBootstrapped_thenContainerIsRunningWithNoExceptions() {
-        assertThat(cassandra.isRunning()).isTrue();
-    }
-
-}
-
-class CarRepositoryLiveTest extends CassandraSimpleIntegrationTest {
-
-    @Autowired
-    private CarRepository carRepository;
-
-    @Test
-    void givenValidCarRecord_whenSavingIt_thenRecordIsSaved() {
-        final UUID carId = UUIDs.timeBased();
-        final Car newCar = new Car(carId, "Nissan", "Qashqai", 2018);
-
-        carRepository.save(newCar);
-
-        final List<Car> savedCars = carRepository.findAllById(List.of(carId));
-        assertThat(savedCars.get(0)).isEqualTo(newCar);
-    }
-
-    @Test
-    void givenExistingCarRecord_whenUpdatingIt_thenRecordIsUpdated() {
-        final UUID carId = UUIDs.timeBased();
-        final Car existingCar = carRepository.save(new Car(carId, "Nissan", "Qashqai", 2018));
-
-        existingCar.setModel("X-Trail");
-        carRepository.save(existingCar);
-
-        final List<Car> savedCars = carRepository.findAllById(List.of(carId));
-        assertThat(savedCars.get(0).getModel()).isEqualTo("X-Trail");
-    }
-
-    @Test
-    void givenExistingCarRecord_whenDeletingIt_thenRecordIsDeleted() {
-        final UUID carId = UUIDs.timeBased();
-        final Car existingCar = carRepository.save(new Car(carId, "Nissan", "Qashqai", 2018));
-
-        carRepository.delete(existingCar);
-
-        final List<Car> savedCars = carRepository.findAllById(List.of(carId));
-        assertThat(savedCars).isEmpty();
     }
 
 }
